TF-972: Gripper cannot deadlock during a given goal.
Added fixed limit timeouts to each of the gripper commands. A couple commands
sent to the gripper waited indefinitely for a given set of register results. In
these cases if the Robotiq failed to process the response correctly the goal
would hang.

Based on TF-853, I'm inclined to believe that there maybe issues in Robotiq's
gripper software in which state doesn't update correctly, but failed to
reproduce the failure reported.