TF-853: Gripper cannot hang in a reset()
Added fixed limit timeouts to each of the grip commands. A couple commands
waited indefinitely for a given set of register results. In these cases the
reset would never complete.

It is possible to send the correct command to the gripper and never receive the
expected result. It seems to be an issue on the Robotiq side.

As a concrete example I observed the gripper fail to complete the activate
command (rACT=1). The queried status from the Robotiq gripper was:

gACT:1, gGTO:0, gSTA:0, gOBJ:0, gFLT:0, gPR:0, gPO:3, gCU:0

According to the documentation on Robotiq's website, a gACT:1 with gSTA:0
implies a fault has occurred.  Unfortunately for us the gFLT register contains
exactly zero useful information about what caused the fault.that waits
indefinitely for a given