Update count
Due to changes made within launch_manager, the Robotiq FT sensor and
the Teachmate struggle when launched simultaneously. This commit
attempts to address that issue by doubling the number of times that the
Robotiq FT sensor will attempt to connect to a serial port.

This change is ultimately a stop-gap pending a complete re-write of this
stack to be more congruent with C++11 standards.